feedback: renamed Carousel component to CarouselInput
