+++ packages/strapi-design-system/src/CarouselInput/Carousel.stories.mdx
@@ -1,4 +1,4 @@
-<!--- Carousel.stories.mdx --->
+<!--- CarouselInput.stories.mdx --->
 
 import { Meta, Story, Canvas } from '@storybook/addon-docs/blocks';
 import { ArgsTable } from '@storybook/addon-docs';
@@ -7,16 +7,16 @@ import Pencil from '@strapi/icons/Pencil';
 import Play from '@strapi/icons/Play';
 import Trash from '@strapi/icons/Trash';
 import Plus from '@strapi/icons/Plus';
-import { Carousel } from './Carousel';
+import { CarouselInput } from './CarouselInput';
 import { CarouselSlide } from './CarouselSlide';
 import { CarouselImage } from './CarouselImage';
 import { CarouselActions } from './CarouselActions';
 import { IconButton } from '../IconButton';
 import firstPathImg from './story-assets/first.jpg';
 
-<Meta title="Design System/Components/Carousel" component={Carousel} />
+<Meta title="Design System/Components/CarouselInput" component={CarouselInput} />
 
-# Carousel
+# CarouselInput
 
 Carousels are slideshow components for navigating through a set of assets.
 
@@ -24,22 +24,22 @@ Carousels are slideshow components for navigating through a set of assets.
 
 - Carousels should be mostly used in forms.
 - Carousels can also be used to display only one asset.
-- The end of a Carousel label should indicate the amount of included assets.
-- A set of actions is possible via Icon Buttons within a Carousel.
+- The end of a CarouselInput label should indicate the amount of included assets.
+- A set of actions is possible via Icon Buttons within a CarouselInput.
 - Carousels should at least give the possibility to remove an asset from it.
 - Carousels don't have any caption but can have a description line.
 
-[View source](https://github.com/strapi/design-system/tree/develop/packages/strapi-design-system/src/Carousel)
+[View source](https://github.com/strapi/design-system/tree/develop/packages/strapi-design-system/src/CarouselInput)
 
 ## Imports
 
 ```js
-import { Carousel, CarouselSlide, CarouselImage, CarouselActions } from '@strapi/design-system/Carousel';
+import { CarouselInput, CarouselSlide, CarouselImage, CarouselActions } from '@strapi/design-system/CarouselInput';
 ```
 
 ## Usage
 
-A Carousel is mostly used in forms. Thanks to their navigational arrows they allow to easily switch from an asset to another. It has an infinite loop. A set of actions is possible via Icon Buttons within a Carousel.
+A CarouselInput is mostly used in forms. Thanks to their navigational arrows they allow to easily switch from an asset to another. It has an infinite loop. A set of actions is possible via Icon Buttons within a CarouselInput.
 
 <Canvas>
   <Story name="base">
@@ -52,7 +52,7 @@ A Carousel is mostly used in forms. Thanks to their navigational arrows they all
         setSelectedIndex((current) => (current > 0 ? current - 1 : 2));
       };
       return (
-        <Carousel
+        <CarouselInput
           label={`Carousel of numbers (${selectedIndex + 1}/3)`}
           selectedSlide={selectedIndex}
           secondaryLabel="/packages/strapi-design-system/src/Carousel/story-assets/first.jpg"
@@ -80,14 +80,14 @@ A Carousel is mostly used in forms. Thanks to their navigational arrows they all
           <CarouselSlide label="3 of 3 slides">
             <CarouselImage src={firstPathImg} alt="third" />
           </CarouselSlide>
-        </Carousel>
+        </CarouselInput>
       );
     }}
   </Story>
   <Story name="one slide only">
     {() => {
       return (
-        <Carousel
+        <CarouselInput
           label={`Carousel of numbers 1/1)`}
           selectedSlide={0}
           previousLabel="Previous slide"
@@ -108,7 +108,7 @@ A Carousel is mostly used in forms. Thanks to their navigational arrows they all
           <CarouselSlide label="1 of 1 slides">
             <CarouselImage src={firstPathImg} alt="First" />
           </CarouselSlide>
-        </Carousel>
+        </CarouselInput>
       );
     }}
   </Story>
@@ -116,4 +116,4 @@ A Carousel is mostly used in forms. Thanks to their navigational arrows they all
 
 ## Props
 
-<ArgsTable of={Carousel} />
+<ArgsTable of={CarouselInput} />
