+++ packages/strapi-design-system/src/CarouselInput/CarouselActions.js
