+++ packages/strapi-design-system/src/CarouselInput/CarouselImage.js
