+++ packages/strapi-design-system/src/CarouselInput/CarouselSlide.js
