+++ packages/strapi-design-system/src/CarouselInput/__tests__/Carousel.e2e.js
