+++ packages/strapi-design-system/src/CarouselInput/index.js
