+++ packages/strapi-design-system/src/CarouselInput/story-assets/first.jpg
